new set

+++ mod-downloader/mod.py
@@ -145,10 +145,10 @@ class Mod:
         self.depend_on_str = depend_on_str
         self.local_dir = os.path.join(mod_file_dir, self.category)
         self.local_path = os.path.join(mod_file_dir, self.category, self.filename)
-        if state != 'inactive' or state == "download":
-            self.link = f'https://{self.domain}.forgecdn.net/files/{link_id(mod_id)}{self.filename}'
-            if os.path.exists(self.local_dir) and os.path.exists(self.local_path):
-                self.download()
+        if self.state != 'inactive' or self.state == "download":
+            self.link = f'https://{self.domain}.forgecdn.net/files/{link_id(self.mod_id)}{self.filename}'
+        else:
+            self.link = None
 
     def depend_on_to_str(self) -> list[str]:
         data = [dep.mod_id for dep in self.depend_on]
@@ -179,19 +179,26 @@ class Mod:
 
     def copy2(self, path: str):
         dest = os.path.join(path, self.filename)
-        if os.path.exists(self.local_dir) and os.path.exists(self.local_path):
-            if not os.path.exists(dest) and self.state == "install":
-                shutil.copyfile(self.local_path, dest)
-                print(f'Copying: {self.local_path} -> {dest}')
-            if self.depend_on is not None and len(self.depend_on) > 0:
-                for dep in self.depend_on:
-                    dep.copy2(path)
-        else:
-            print(f"Local mod not found: {self.local_path}")
+        if not os.path.exists(self.local_path):
+            create_dir(self.local_dir)
+            if self.link is not None:
+                if os.path.exists(self.local_dir) and not os.path.exists(self.local_path):
+                    self.download()
+        
+        if not os.path.exists(dest) and self.state == "install" and os.path.exists(self.local_path):
+            shutil.copyfile(self.local_path, dest)
+            print(f'Copying: {self.local_path} -> {dest}')
+        if self.depend_on is not None and len(self.depend_on) > 0:
+            for dep in self.depend_on:
+                dep.copy2(path)
+            
 
     def remove(self):
         print(f"Removing: {self.local_path}")
         os.remove(os.path.join(mod_file_dir, self.local_path))
+        if self.depend_on is not None and len(self.depend_on) > 0:
+            for dep in self.depend_on:
+                dep.remove()
 
 
 class ModPack:
@@ -200,11 +207,8 @@ class ModPack:
         return cls(pack_json['name'], pack_json['display_name'], pack_json['description'], pack_content)
 
     @classmethod
-    def create_pack(cls) -> 'ModPack':
-        name = input('Enter a pack name: ')
-        display_name = input('Enter a display name: ')
-        description = input('Enter a description: ')
-        return cls(name, display_name, description, [])
+    def create_pack(cls, name, display_name, description, mods) -> 'ModPack':
+        return cls(name, display_name, description, mods)
 
     @classmethod
     def init_pack(cls, existing_pack: 'ModPack') -> 'ModPack':
@@ -270,8 +274,8 @@ class ModManager:
                     mod = Mod.load_from_json(mod_data, category)
                     if mod:
                         self.mod_list.append(mod)
-                        if not os.path.exists(mod.local_path) and mod.state != "inactive":
-                            mod.download()
+                        # if not os.path.exists(mod.local_path) and mod.state != "inactive":
+                        #     mod.download()
                         if category not in self.mod_categories:
                             self.mod_categories.append(category)
